Complete the merge by combining env and env_var.go
