+++ model/env.go
@@ -1,6 +1,8 @@
 package model
 
 import (
+	"encoding/json"
+
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 )
@@ -48,4 +50,26 @@ func (em *EnvVarMap) ToEnvList() []corev1.EnvVar {
 	}
 
 	return envList
+}
+
+// ToJSON converts the EnvVarMap to a JSON object represented as a
+// []byte
+func (e *EnvVarMap) ToJSON() ([]byte, error) {
+	b, err := json.Marshal(e)
+	if err != nil {
+		return []byte(nil), err
+	}
+
+	return b, nil
+}
+
+// EnvVarFromJSON creates a EnvVarMap from the JSON represented as a
+// []byte
+func EnvVarFromJSON(raw []byte) (*EnvVarMap, error) {
+	e := &EnvVarMap{}
+	err := json.Unmarshal(raw, e)
+	if err != nil {
+		return nil, err
+	}
+	return e, nil
 }
